Refactored header styles.
