+++ example/index.html
@@ -4,7 +4,7 @@
     <title>Apollon Editor</title>
     <meta charset="utf-8" />
 
-    <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/css/bootstrap-reboot.min.css" />
+    <!-- <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/twitter-bootstrap/4.3.1/css/bootstrap-reboot.min.css" /> -->
     <link rel="stylesheet" href="styles.css" />
   </head>
   <body>
