--- src/components/controls/header.ts
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const Header = styled.strong`
-  display: block;
-  font-family: HelveticaNeue-Medium, sans-serif;
-  font-weight: normal;
-  margin-bottom: 7px;
-`;
