+++ src/components/controls/typography/typography.tsx
@@ -1,0 +1,7 @@
+import React, { AllHTMLAttributes } from 'react';
+import { Omit } from 'react-redux';
+import { Typography } from './typography-styles';
+
+export const Header = (props: Omit<AllHTMLAttributes<HTMLElement>, 'as'>) => (
+  <Typography variant="header" as="h1" {...props} />
+);
