+++ src/packages/activity-diagram/activity-control-flow/activity-control-flow-popup.tsx
@@ -3,8 +3,8 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import styled from 'styled-components';
 import { FlipIcon } from '../../../components/controls/flip-icon';
-import { Header } from '../../../components/controls/header';
 import { Textfield } from '../../../components/controls/textfield/textfield';
+import { Header } from '../../../components/controls/typography/typography';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
 import { ElementRepository } from '../../../services/element/element-repository';
