+++ src/packages/class-diagram/classifier/classifier-popup.tsx
@@ -4,10 +4,10 @@ import { compose } from 'redux';
 import styled from 'styled-components';
 import { ClassElementType } from '..';
 import { Divider } from '../../../components/controls/divider/divider';
-import { Header } from '../../../components/controls/header';
 import { Switch } from '../../../components/controls/switch/switch';
 import { Textfield } from '../../../components/controls/textfield/textfield';
 import { TrashCanIcon } from '../../../components/controls/trashcan';
+import { Header } from '../../../components/controls/typography/typography';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
 import { ModelState } from '../../../components/store/model-state';
