+++ src/packages/communication-diagram/communication-link/communication-link-popup.tsx
@@ -4,9 +4,9 @@ import { compose } from 'redux';
 import styled from 'styled-components';
 import { CommunicationMessage } from '..';
 import { Divider } from '../../../components/controls/divider/divider';
-import { Header } from '../../../components/controls/header';
 import { Textfield } from '../../../components/controls/textfield/textfield';
 import { TrashCanIcon } from '../../../components/controls/trashcan';
+import { Header } from '../../../components/controls/typography/typography';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
 import { ModelState } from '../../../components/store/model-state';
