+++ src/packages/object-diagram/object-name/object-name-popup.tsx
@@ -3,9 +3,9 @@ import { connect } from 'react-redux';
 import { compose } from 'redux';
 import styled from 'styled-components';
 import { Divider } from '../../../components/controls/divider/divider';
-import { Header } from '../../../components/controls/header';
 import { Textfield } from '../../../components/controls/textfield/textfield';
 import { TrashCanIcon } from '../../../components/controls/trashcan';
+import { Header } from '../../../components/controls/typography/typography';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
 import { Element } from '../../../services/element/element';
