+++ src/packages/use-case-diagram/use-case-association/use-case-association-popup.tsx
@@ -6,8 +6,8 @@ import { UseCaseRelationshipType } from '..';
 import { Divider } from '../../../components/controls/divider/divider';
 import { Dropdown } from '../../../components/controls/dropdown/dropdown';
 import { FlipIcon } from '../../../components/controls/flip-icon';
-import { Header } from '../../../components/controls/header';
 import { Textfield } from '../../../components/controls/textfield/textfield';
+import { Header } from '../../../components/controls/typography/typography';
 import { I18nContext } from '../../../components/i18n/i18n-context';
 import { localized } from '../../../components/i18n/localized';
 import { Element } from '../../../services/element/element';
